IMPORTANT COMMIT. Implement dynamic level loader.
Made changes in the scene controller so that there is a list of level names that the _update_scene and whatever else functions pull from. This seems to properly move one by one through the array, going level by level.

From here on, level naming conventions must match what is in the LEVEL_LIST.

We should also make sure that the next_level_counter can't count out of bounds, although this shouldn't happen as long as we don't have more level ends than there are levels.*

*In the current state, the level end field in Level3 would crash the game, as it would try to find a nonexistent Level4.