Watched Luke's video about making enemies
Made a basic enemy that walks forward and can be stomped on.

Does not yet have functionality to turn around when hitting a wall.

My plan is to maybe make it a drone security robot thing. Unkillable, when touched sets off an alarm. Maybe also give it a view cone to avoid as well? Overall shouldn't be too difficult.