Create spritesheet for robot enemy
Made the robot's sprite sheet and implemented that, so the robot walks, or I guess drives, with a little animation.

In the future, I need to give it raycasting so that it turns around when hitting a wall (and maybe also ledges?)

I also plan to give it a collision on top of its head, so that the player may stand on top of it, to sort of "catch a ride." Could be used in a level where the robot doesn't set off lasers, so the player uses it to cross an impossible jump.