Add raycast to enemy to detect walls, and ride on enemy heads
Robot enemies now have a raycast that detects world collisions, telling the robot to turn around.

Also, I made it possible to ride on top of a robot by just getting rid of the goomba stomp hitbox functionality on top of it, and changing one thing in collision for Clepto.

Godot handled all the physics and is_on_floor stuff, thank goodness.