Add hatch animation and fix level loader
Added in the spritesheet for the end of level hatch, and made it so that when the player touches the area under it, the function waits until the hatch closing animation has finished before moving to the next level. Not to toot my own horn, but its not a bad animation. :)

Also, the level loader has been buggy, so I checked it out and realized that right now it was set up to be TOLD what level to go to, rather than track it itself. Specifically, EndAreas would call it and could only tell it to go to one specific level until I moved that responsibility to the scene_controller.

Level loading is basically flawless now, just need to add the exact name of the level scene to the LEVEL_LIST array and keep track of where we are.