Add walk, crouch, jump, walljump SFX
Implemented these SFX. May still need to fix the volume levels of some of them, specifically the WalkAudio.

WALK_AUDIO had a weird bug where in the first frame, it would be null and crash the game. Fixed it with a simple check for if it is null, don't try to run stop() on it. Don't know why it was even null, but oh well.