Add hatch close sound
Added a sound for the hatch closing.
Tried to implement it in a way where it plays the sound on an exact frame of the animation, but that was too weird. Internet wasn't helpful.
Instead, I just used another timer to play it at the right time. Worked well.

Also tweaked clepto walk decibel to be louder, and lowered the camera on level 4 so you can see the robots you're jumping on.